Added support for custom bootloader templates
The support for a custom grub bootloader template allows to
specify a template file which will be used instead of the one
provided with kiwi. A static bootloader template to create
the grub config file is only used in kiwi if the native
method via the grub mkconfig toolchain does not work properly.
As of today this is only the case for live and install ISO
images. Thus the value is expected to be effective for
the oem and iso image types only

This is related to Issue #1970

!!! WIP , no merge !!!